refac: comment controller and service

+++ src/main/java/com/luizgmelo/conduit/services/CommentService.java
@@ -1,6 +1,11 @@
 package com.luizgmelo.conduit.services;
 
+import com.luizgmelo.conduit.dtos.CommentDTO;
 import com.luizgmelo.conduit.dtos.CommentRequestDto;
+import com.luizgmelo.conduit.dtos.CommentResponseDTO;
+import com.luizgmelo.conduit.dtos.MultipleCommentResponseDTO;
+import com.luizgmelo.conduit.exceptions.CommentNotFoundException;
+import com.luizgmelo.conduit.exceptions.OperationNotAllowedException;
 import com.luizgmelo.conduit.models.Article;
 import com.luizgmelo.conduit.models.Comment;
 import com.luizgmelo.conduit.models.User;
@@ -11,39 +16,51 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
 public class CommentService {
 
   private final CommentRepository commentRepository;
-  private final UserService userService;
+  private final ArticleService articleService;
 
-  public CommentService(CommentRepository commentRepository, UserService userService) {
+  public CommentService(CommentRepository commentRepository, ArticleService articleService) {
     this.commentRepository = commentRepository;
-    this.userService = userService;
+    this.articleService = articleService;
   }
 
   public Comment getCommentById(UUID commentId) {
-    return commentRepository.findById(commentId).orElse(null);
+    return commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
   }
 
-
-  public Page<Comment> getAllComments(Article article, int page, int size) {
+  public MultipleCommentResponseDTO getAllComments(String slug, int page, int size) {
+    Article article = articleService.getArticleBySlug(slug);
     Pageable pageable = PageRequest.of(page, size, Sort.by("createdAt").descending());
-    return commentRepository.findByCommentFromId(article.getId(), pageable);
+    Page<Comment> comments = commentRepository.findByCommentFromId(article.getId(), pageable);
+    List<CommentDTO> commentDTOList = comments.getContent().stream().map(CommentDTO::fromComment).toList();
+    return new MultipleCommentResponseDTO(commentDTOList);
   }
 
-  public Comment createComment(Article article, CommentRequestDto commentDto) {
+  public CommentResponseDTO createComment(String slug, User user,CommentRequestDto commentDto) {
+    Article article = articleService.getArticleBySlug(slug);
+
     String comment = commentDto.comment().body();
-    User author = userService.getAuthenticatedUser();
 
-    Comment newComment = new Comment(comment, article, author);
+    Comment newComment = new Comment(comment, article, user);
 
-    return commentRepository.save(newComment);
+    Comment commentSaved = commentRepository.save(newComment);
+    return CommentResponseDTO.fromComment(commentSaved);
   }
 
-  public void deleteComment(String slug, Comment comment) {
+  public void deleteComment(User user, String slug, UUID commentId) {
+    Article article = articleService.getArticleBySlug(slug);
+    Comment comment = this.getCommentById(commentId);
+
+    if (!user.getEmail().equalsIgnoreCase(article.getAuthor().getEmail())) {
+      throw new OperationNotAllowedException();
+    }
+
     if (comment.getCommentFrom().getSlug().equals(slug)) {
       commentRepository.delete(comment);
     }
